# services/client-api/api_specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaxEase Backend API",
    "description": "Comprehensive API for Canadian tax preparation and filing system with Flutter mobile app integration",
    "version": "1.0.0",
    "contact": {
      "name": "TaxEase Support",
      "email": "support@taxease.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.taxease.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TaxForm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tax_year": {
            "type": "integer",
            "example": 2024
          },
          "status": {
            "type": "string",
            "enum": ["draft", "calculated", "submitted", "accepted"],
            "example": "calculated"
          },
          "sin": {
            "type": "string",
            "example": "123456789"
          },
          "total_income": {
            "type": "number",
            "format": "decimal",
            "example": 77500.00
          },
          "total_tax": {
            "type": "number",
            "format": "decimal",
            "example": 15000.00
          },
          "refund_or_owing": {
            "type": "number",
            "format": "decimal",
            "example": -2500.00
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/tax/forms": {
      "get": {
        "summary": "List Tax Forms",
        "description": "Get all tax forms for the current user",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 10},
            "description": "Number of forms to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {"type": "integer", "default": 0},
            "description": "Number of forms to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tax forms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "forms": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/TaxForm"}
                    },
                    "total": {"type": "integer", "example": 1},
                    "limit": {"type": "integer", "example": 10},
                    "offset": {"type": "integer", "example": 0}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Tax Form",
        "description": "Create a new T1 personal tax form",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tax_year", "sin"],
                "properties": {
                  "tax_year": {"type": "integer", "example": 2024},
                  "sin": {"type": "string", "example": "123456789"},
                  "date_of_birth": {"type": "string", "format": "date", "example": "1990-01-15"},
                  "marital_status": {"type": "string", "enum": ["single", "married", "common_law", "divorced"], "example": "single"},
                  "employment_income": {"type": "number", "example": 75000.00},
                  "self_employment_income": {"type": "number", "example": 0.00},
                  "investment_income": {"type": "number", "example": 2500.00},
                  "rrsp_contributions": {"type": "number", "example": 5000.00},
                  "charitable_donations": {"type": "number", "example": 500.00}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tax form created successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TaxForm"}
              }
            }
          }
        }
      }
    },
    "/tax/forms/{form_id}": {
      "get": {
        "summary": "Get Tax Form",
        "description": "Get details of a specific tax form",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {
          "200": {
            "description": "Tax form details",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TaxForm"}
              }
            }
          },
          "404": {
            "description": "Tax form not found"
          }
        }
      },
      "put": {
        "summary": "Update Tax Form",
        "description": "Update an existing tax form",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employment_income": {"type": "number", "example": 80000.00},
                  "charitable_donations": {"type": "number", "example": 1000.00}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tax form updated successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TaxForm"}
              }
            }
          }
        }
      }
    },
...
  },
...
}